Moved references to the bot into a decorator class
