+++ lib/farmbot-pi.rb
@@ -4,6 +4,7 @@ require_relative 'messaging/credentials'
 require_relative 'messaging/message_handler'
 require_relative 'chores/chore_runner'
 require_relative 'models/status_storage.rb'
+require_relative 'bot_decorator'
 require 'pry'
 
 ActiveRecord::Base.establish_connection(
@@ -15,10 +16,12 @@ class FarmBotPi
   attr_accessor :mesh, :bot, :credentials, :handler, :runner, :status_storage
 
   def initialize(bot: select_correct_bot)
-    @credentials = FBPi::Credentials.new
-    @mesh        = EM::MeshRuby.new(@credentials.uuid, @credentials.token, 'ws://mesh.farmbot.it')
-    @bot         = bot
+    @credentials    = FBPi::Credentials.new
+    @mesh           = EM::MeshRuby.new(@credentials.uuid,
+                                       @credentials.token,
+                                       'ws://mesh.farmbot.it')
     @status_storage = FBPi::StatusStorage.new("bot_status.pstore")
+    @bot            = FBPi::BotDecorator.build(bot, @status_storage, @mesh)
   end
 
   def select_correct_bot
@@ -33,35 +36,17 @@ class FarmBotPi
 
   def start
     EM.run do
-      load_previous_state
       mesh.toggle_debug!
       mesh.connect
       FB::ArduinoEventMachine.connect(bot)
       start_chore_runner
       broadcast_status
       mesh.onmessage { |msg| meshmessage(msg) }
-      bot.onmessage  { |msg| botmessage(msg) }
-      bot.onchange   { |msg| diffmessage(msg) }
-      bot.onclose    { |msg| close(msg) }
+      bot.bootstrap
       this = self # Binding of caller makes me sad :(
       mesh.socket.on(:ready) do
         this.mesh.data(this.status_storage.to_h.merge(log: "online"))
       end
-    end
-  end
-
-  def load_previous_state
-    previous_states = status_storage.to_h
-    bot.status.transaction do |status|
-      previous_states.each { |k,v| status[k] = v }
-    end
-  end
-
-  def botmessage(msg)
-    if msg.name == :idle
-      print '.'
-    else
-      bot.log "BOT MSG: #{msg.name} #{msg.to_s}"
     end
   end
 
@@ -82,16 +67,5 @@ class FarmBotPi
                                        payload: {})
       FBPi::ReadStatusController.new(null_msg, bot, mesh).call
     end
-  end
-
-  def diffmessage(diff)
-    @status_storage.update_attributes(diff)
-    bot.log "BOT DIF: #{diff}" unless diff.keys == [:BUSY]
-  end
-
-  def close(_args)
-    @status_storage.update_attributes(bot.status.to_h)
-    @mesh.data @status_storage.to_h.merge(log: "offline")
-    EM.stop
   end
 end
